requirements.txt
scrapy
slugify
bs4
mysql-connector-python